fix potential type mismatch between format specifier and argument
Cast argument of type rlim_t to unsigned long long to match the %llu
format specifier on platforms where this might not be the case.

+++ gotwebd/sockets.c
@@ -709,7 +709,8 @@ sockets_rlimit(int maxfd)
 
 	if (getrlimit(RLIMIT_NOFILE, &rl) == -1)
 		fatal("%s: failed to get resource limit", __func__);
-	log_debug("%s: max open files %llu", __func__, rl.rlim_max);
+	log_debug("%s: max open files %llu", __func__,
+	    (unsigned long long)rl.rlim_max);
 
 	/*
 	 * Allow the maximum number of open file descriptors for this
